refactor(zk): pure cleanup/reorganize of groth16 module
